removed duplicated route
